requirements.txt
langchain
openai==0.27.6
python-dotenv==1.0.0
apikey==0.2.4 
pymongo==4.3.3
pylint==2.17.4
fuzzywuzzy==0.18.0 
sentence_transformers==2.2.2
streamlit==1.27.0 
PyPDF2==3.0.1 
openpyxl==3.1.2 
faiss-cpu==1.7.4 
tiktoken
geopy==2.4.0 
folium==0.14.0 
streamlit-folium
pinecone-client
tabulate
transformers
InstructorEmbedding
cohere
pyht
numpy==1.26.0
facebook-sdk
python-Levenshtein
selenium
streamlit
seleniumbase
webdriver-manager
undetected_chromedriver==3.5.3
transformers==4.35.2
elevenlabs
ffmpeg
ffmpeg-python